+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * fix problem were await affect quoted promise in macros (e.g. `begin`) [#153](https://github.com/jcubic/lips/issues/153)
 * typecheck second argument to `set-obj!`
 * fix `case` macro (use implementation from R7RS Errata)
+* fix async constructor in `define-class`
 
 ## 1.0.0-beta.12
 ### Breaking